+++ server/WodssServer/src/main/java/ch/fhnw/wodss/service/AttachmentService.java
@@ -1,8 +1,12 @@
 package ch.fhnw.wodss.service;
 
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
 import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,7 +33,10 @@ public class AttachmentService {
 	
 	@Transactional
 	public void deleteAttachment(Attachment attachment){
-		attachmentRepository.delete(attachment);
+		boolean success = deleteAttachmentFromFileSystem(attachment);
+		if(success){
+			attachmentRepository.delete(attachment);
+		}
 	}
 	
 	/**
@@ -39,11 +46,43 @@ public class AttachmentService {
 	 * @return success
 	 */
 	public boolean saveAttachmentToFileSystem(Attachment attachment, MultipartFile file) {
+		try {
+			return saveAttachmentToFileSystem(attachment, file.getBytes());
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+		return false;
+	}
+	
+	/**
+	 * Saved the uploaded file to file system
+	 * @param attachment the attachment object to retrieve path from uuid
+	 * @param file the file to save.
+	 * @return success
+	 */
+	public boolean saveAttachmentToFileSystem(Attachment attachment, File file) {
+		try {
+			Path path = Paths.get(file.toURI());
+			byte[] bytes = Files.readAllBytes(path);
+			return saveAttachmentToFileSystem(attachment, bytes);
+		} catch (IOException e) {
+			e.printStackTrace();
+		}
+		return false;
+	}
+	
+	/**
+	 * Saved the uploaded file to file system
+	 * @param attachment the attachment object to retrieve path from uuid
+	 * @param bytes the file in bytes to save.
+	 * @return success
+	 */
+	public boolean saveAttachmentToFileSystem(Attachment attachment, byte[] bytes) {
 		FileOutputStream fos = null;
 		try {
 			FileUtils.forceMkdir(attachment.getFile().getParentFile());
 			fos = new FileOutputStream(attachment.getFile());
-			fos.write(file.getBytes());
+			fos.write(bytes);
 			return true;
 		} catch (FileNotFoundException e) {
 			e.printStackTrace();
